Moving FullScreenAction from app-display-runtime to core-ui Creating authorization entry for the FullScreenAction
