+++ app/display/runtime/src/main/java/org/csstudio/display/builder/runtime/Messages.java
@@ -12,8 +12,6 @@ import org.phoebus.framework.nls.NLS;
 public class Messages
 {
     // Keep in alphabetical order and aligned with messages.properties
-    public static String EnterFullscreen;
-    public static String ExitFullscreen;
     public static String NavigateBack_TT;
     public static String NavigateForward_TT;
     public static String OpenDataBrowser;
