+++ app/display/runtime/src/main/resources/org/csstudio/display/builder/runtime/messages.properties
@@ -1,5 +1,3 @@
-EnterFullscreen=Full-screen
-ExitFullscreen=Exit Full-screen
 NavigateBack_TT=Open previous display
 NavigateForward_TT=Open next display
 OpenDataBrowser=Open Data Browser
