+++ core/security/src/main/resources/authorization.conf
@@ -11,6 +11,9 @@
 # Anybody can lock a dock pane, i.e. set it to 'fixed'
 lock_ui = .*
 
+# Anybody can enter and exit full screen
+full_screen = .*
+
 # Anybody can disable save/restore for a pvs file.
 disable_save_restore = .*
