+++ core/ui/src/main/resources/icons/exitfullscreen.png
