+++ core/ui/src/main/resources/icons/fullscreen.png
