+++ core/ui/src/main/resources/org/phoebus/ui/javafx/messages.properties
@@ -9,6 +9,8 @@ Clipboard=Clipboard Image
 ClipboardTooltip=Add an image from the clipboard
 CSSWindow=CSS Window
 CSSWindowTooltip=Add a capture of the application window
+EnterFullScreen=Full-screen
+ExitFullScreen=Exit Full-screen
 Files=Files
 Images=Images
 ImagesTitle=Images:
